+++ src/DataContainer/Content.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 /**
  * SMARTGEAR for Contao Open Source CMS
- * Copyright (c) 2015-2022 Web ex Machina
+ * Copyright (c) 2015-2023 Web ex Machina
  *
  * @category ContaoBundle
  * @package  Web-Ex-Machina/contao-smartgear
@@ -15,12 +15,16 @@ declare(strict_types=1);
 namespace WEM\SmartgearBundle\DataContainer;
 
 use Contao\Backend;
+use Contao\Config;
 use Contao\ContentModel;
 use Contao\CoreBundle\Exception\AccessDeniedException;
 use Contao\DataContainer;
+use Contao\Date;
 use Contao\Image;
 use Contao\Input;
 use Contao\System;
+use DateInterval;
+use DateTime;
 use Exception;
 use tl_content;
 use tl_content_calendar;